Fixes after CR
