+++ src/SprykerEco/Zed/Unzer/Business/ApiAdapter/Mapper/UnzerAuthorizePaymentMapper.php
@@ -9,18 +9,16 @@ namespace SprykerEco\Zed\Unzer\Business\ApiAdapter\Mapper;
 
 use Generated\Shared\Transfer\UnzerApiAuthorizeRequestTransfer;
 use Generated\Shared\Transfer\UnzerApiAuthorizeResponseTransfer;
-use Generated\Shared\Transfer\UnzerApiErrorResponseTransfer;
 use Generated\Shared\Transfer\UnzerApiMarketplaceAuthorizeRequestTransfer;
 use Generated\Shared\Transfer\UnzerApiMarketplaceAuthorizeResponseTransfer;
 use Generated\Shared\Transfer\UnzerBasketTransfer;
 use Generated\Shared\Transfer\UnzerCustomerTransfer;
-use Generated\Shared\Transfer\UnzerPaymentErrorTransfer;
 use Generated\Shared\Transfer\UnzerPaymentResourceTransfer;
 use Generated\Shared\Transfer\UnzerPaymentTransfer;
 use SprykerEco\Zed\Unzer\UnzerConfig;
 use SprykerEco\Zed\Unzer\UnzerConstants;
 
-class UnzerAuthorizePaymentMapper implements UnzerAuthorizePaymentMapperInterface
+class UnzerAuthorizePaymentMapper extends UnzerErrorMapper implements UnzerAuthorizePaymentMapperInterface
 {
     /**
      * @var \SprykerEco\Zed\Unzer\UnzerConfig
@@ -110,26 +108,5 @@ class UnzerAuthorizePaymentMapper implements UnzerAuthorizePaymentMapperInterfac
             ->setCustomer((new UnzerCustomerTransfer())->setId($unzerApiAuthorizeResponseTransfer->getCustomerId()))
             ->setPaymentResource((new UnzerPaymentResourceTransfer())->setId($unzerApiAuthorizeResponseTransfer->getTypeId()))
             ->setRedirectUrl($unzerApiAuthorizeResponseTransfer->getRedirectUrl());
-    }
-
-    /**
-     * @param \Generated\Shared\Transfer\UnzerApiErrorResponseTransfer $unzerApiErrorResponseTransfer
-     * @param \Generated\Shared\Transfer\UnzerPaymentTransfer $unzerPaymentTransfer
-     *
-     * @return \Generated\Shared\Transfer\UnzerPaymentTransfer
-     */
-    public function mapUnzerApiErrorResponseTransferToUnzerPaymentTransfer(
-        UnzerApiErrorResponseTransfer $unzerApiErrorResponseTransfer,
-        UnzerPaymentTransfer $unzerPaymentTransfer
-    ): UnzerPaymentTransfer {
-        foreach ($unzerApiErrorResponseTransfer->getErrors() as $unzerApiResponseErrorTransfer) {
-            $unzerPaymentTransfer->addError(
-                (new UnzerPaymentErrorTransfer())
-                    ->setMessage($unzerApiResponseErrorTransfer->getCustomerMessage())
-                    ->setErrorCode((int)$unzerApiResponseErrorTransfer->getCode()),
-            );
-        }
-
-        return $unzerPaymentTransfer;
     }
 }
