+++ src/SprykerEco/Zed/Unzer/Business/ApiAdapter/Mapper/UnzerAuthorizePaymentMapperInterface.php
@@ -9,7 +9,6 @@ namespace SprykerEco\Zed\Unzer\Business\ApiAdapter\Mapper;
 
 use Generated\Shared\Transfer\UnzerApiAuthorizeRequestTransfer;
 use Generated\Shared\Transfer\UnzerApiAuthorizeResponseTransfer;
-use Generated\Shared\Transfer\UnzerApiErrorResponseTransfer;
 use Generated\Shared\Transfer\UnzerApiMarketplaceAuthorizeRequestTransfer;
 use Generated\Shared\Transfer\UnzerApiMarketplaceAuthorizeResponseTransfer;
 use Generated\Shared\Transfer\UnzerPaymentTransfer;
@@ -57,17 +56,6 @@ interface UnzerAuthorizePaymentMapperInterface
      */
     public function mapUnzerApiAuthorizeResponseTransferToUnzerPaymentTransfer(
         UnzerApiAuthorizeResponseTransfer $unzerApiAuthorizeResponseTransfer,
-        UnzerPaymentTransfer $unzerPaymentTransfer
-    ): UnzerPaymentTransfer;
-
-    /**
-     * @param \Generated\Shared\Transfer\UnzerApiErrorResponseTransfer $unzerApiErrorResponseTransfer
-     * @param \Generated\Shared\Transfer\UnzerPaymentTransfer $unzerPaymentTransfer
-     *
-     * @return \Generated\Shared\Transfer\UnzerPaymentTransfer
-     */
-    public function mapUnzerApiErrorResponseTransferToUnzerPaymentTransfer(
-        UnzerApiErrorResponseTransfer $unzerApiErrorResponseTransfer,
         UnzerPaymentTransfer $unzerPaymentTransfer
     ): UnzerPaymentTransfer;
 }
