+++ src/SprykerEco/Zed/Unzer/Business/ApiAdapter/Mapper/UnzerGetPaymentMapper.php
@@ -8,15 +8,13 @@
 namespace SprykerEco\Zed\Unzer\Business\ApiAdapter\Mapper;
 
 use ArrayObject;
-use Generated\Shared\Transfer\UnzerApiErrorResponseTransfer;
 use Generated\Shared\Transfer\UnzerApiGetPaymentRequestTransfer;
 use Generated\Shared\Transfer\UnzerApiGetPaymentResponseTransfer;
-use Generated\Shared\Transfer\UnzerPaymentErrorTransfer;
 use Generated\Shared\Transfer\UnzerPaymentTransfer;
 use Generated\Shared\Transfer\UnzerTransactionTransfer;
 use SprykerEco\Zed\Unzer\UnzerConstants;
 
-class UnzerGetPaymentMapper implements UnzerGetPaymentMapperInterface
+class UnzerGetPaymentMapper extends UnzerErrorMapper implements UnzerGetPaymentMapperInterface
 {
     /**
      * @param \Generated\Shared\Transfer\UnzerPaymentTransfer $unzerPaymentTransfer
@@ -53,27 +51,6 @@ class UnzerGetPaymentMapper implements UnzerGetPaymentMapperInterface
         $unzerPaymentTransactions = $this->mapUnzerApiGetPaymentResponseTransferToUnzerTransactionTransfers($unzerApiGetPaymentResponseTransfer);
 
         return $unzerPaymentTransfer->setTransactions($unzerPaymentTransactions);
-    }
-
-    /**
-     * @param \Generated\Shared\Transfer\UnzerApiErrorResponseTransfer $unzerApiErrorResponseTransfer
-     * @param \Generated\Shared\Transfer\UnzerPaymentTransfer $unzerPaymentTransfer
-     *
-     * @return \Generated\Shared\Transfer\UnzerPaymentTransfer
-     */
-    public function mapUnzerApiErrorResponseTransferToUnzerPaymentTransfer(
-        UnzerApiErrorResponseTransfer $unzerApiErrorResponseTransfer,
-        UnzerPaymentTransfer $unzerPaymentTransfer
-    ): UnzerPaymentTransfer {
-        foreach ($unzerApiErrorResponseTransfer->getErrors() as $unzerApiResponseErrorTransfer) {
-            $unzerPaymentTransfer->addError(
-                (new UnzerPaymentErrorTransfer())
-                    ->setMessage($unzerApiResponseErrorTransfer->getCustomerMessage())
-                    ->setErrorCode((int)$unzerApiResponseErrorTransfer->getCode()),
-            );
-        }
-
-        return $unzerPaymentTransfer;
     }
 
     /**
